Extract score and score functions to new files
Score is no longer defined in VotingRound, meaning
tests which only verify the score logic no longer need
to depend on VotingRound.

Note that there is some additional mess to clean up now,
since combine() in functions.cpp uses the new score files.
What this means is that any test which depends on functions.cpp
now also needs to depend on score.cpp and score_helpers.cpp to be
able to compile. This will need to be cleaned up later.
