+++ src/score.cpp
@@ -1,0 +1,5 @@
+#include "score.h"
+
+auto operator==(Score const& a, Score const& b) -> bool {
+	return a.item == b.item && a.wins == b.wins && a.losses == b.losses;
+}
