+++ src/test/CMakeLists.txt
@@ -8,6 +8,7 @@ add_executable(${test_name}
 	testing.cpp
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
 	${PROJECT_SOURCE_DIR}/src/print.cpp
+	${PROJECT_SOURCE_DIR}/src/score.cpp
 	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
@@ -16,10 +17,10 @@ set(test_name test_combine_scores)
 add_executable(${test_name}
 	${test_name}.cpp
 	testing.cpp
-	${PROJECT_SOURCE_DIR}/src/functions.cpp
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
 	${PROJECT_SOURCE_DIR}/src/print.cpp
-	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
+	${PROJECT_SOURCE_DIR}/src/score.cpp
+	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
 
@@ -27,10 +28,10 @@ set(test_name test_create_score_table)
 add_executable(${test_name}
 	${test_name}.cpp
 	testing.cpp
-	${PROJECT_SOURCE_DIR}/src/functions.cpp
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
 	${PROJECT_SOURCE_DIR}/src/print.cpp
-	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
+	${PROJECT_SOURCE_DIR}/src/score.cpp
+	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
 
@@ -58,10 +59,10 @@ set(test_name test_generate_score_file_data)
 add_executable(${test_name}
 	${test_name}.cpp
 	testing.cpp
-	${PROJECT_SOURCE_DIR}/src/functions.cpp
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
 	${PROJECT_SOURCE_DIR}/src/print.cpp
-	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
+	${PROJECT_SOURCE_DIR}/src/score.cpp
+	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
 
@@ -79,9 +80,11 @@ set(test_name test_get_active_menu_string)
 add_executable(${test_name}
 	${test_name}.cpp
 	testing.cpp
-	${PROJECT_SOURCE_DIR}/src/functions.cpp
+	${PROJECT_SOURCE_DIR}/src/functions.cpp			# TODO: Remove dependency to this file here
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
 	${PROJECT_SOURCE_DIR}/src/print.cpp
+	${PROJECT_SOURCE_DIR}/src/score.cpp				# TODO: Remove dependency to this file here
+	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp		# TODO: Remove dependency to this file here
 	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
@@ -100,10 +103,8 @@ set(test_name test_load_and_save_file)
 add_executable(${test_name}
 	${test_name}.cpp
 	testing.cpp
-	${PROJECT_SOURCE_DIR}/src/functions.cpp
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
-	${PROJECT_SOURCE_DIR}/src/print.cpp
-	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
+	${PROJECT_SOURCE_DIR}/src/print.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
 
@@ -111,10 +112,10 @@ set(test_name test_parse_scores)
 add_executable(${test_name}
 	${test_name}.cpp
 	testing.cpp
-	${PROJECT_SOURCE_DIR}/src/functions.cpp
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
 	${PROJECT_SOURCE_DIR}/src/print.cpp
-	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
+	${PROJECT_SOURCE_DIR}/src/score.cpp
+	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
 
@@ -138,6 +139,19 @@ add_executable(${test_name}
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
 
+set(test_name test_save_voting_round)
+add_executable(${test_name}
+	${test_name}.cpp
+	testing.cpp
+	${PROJECT_SOURCE_DIR}/src/functions.cpp			# TODO: Remove dependency to this file here
+	${PROJECT_SOURCE_DIR}/src/helpers.cpp
+	${PROJECT_SOURCE_DIR}/src/print.cpp
+	${PROJECT_SOURCE_DIR}/src/score.cpp				# TODO: Remove dependency to this file here
+	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp		# TODO: Remove dependency to this file here
+	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
+target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
+add_test(NAME ${test_name} COMMAND ${test_name})
+
 set(test_name test_shuffle_voting_order)
 add_executable(${test_name}
 	${test_name}.cpp
@@ -155,6 +169,8 @@ add_executable(${test_name}
 	${PROJECT_SOURCE_DIR}/src/functions.cpp
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
 	${PROJECT_SOURCE_DIR}/src/print.cpp
+	${PROJECT_SOURCE_DIR}/src/score.cpp				# TODO: Remove dependency to this file here
+	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp		# TODO: Remove dependency to this file here
 	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
@@ -176,6 +192,8 @@ add_executable(${test_name}
 	${PROJECT_SOURCE_DIR}/src/functions.cpp
 	${PROJECT_SOURCE_DIR}/src/helpers.cpp
 	${PROJECT_SOURCE_DIR}/src/print.cpp
+	${PROJECT_SOURCE_DIR}/src/score.cpp				# TODO: Remove dependency to this file here
+	${PROJECT_SOURCE_DIR}/src/score_helpers.cpp		# TODO: Remove dependency to this file here
 	${PROJECT_SOURCE_DIR}/src/voting_round.cpp)
 target_include_directories(${test_name} PRIVATE ${PROJECT_SOURCE_DIR}/src)
 add_test(NAME ${test_name} COMMAND ${test_name})
