+++ src/test/test_combine_scores.cpp
@@ -1,5 +1,5 @@
-#include "functions.h"
 #include "testing.h"
+#include "score_helpers.h"
 
 namespace
 {
